[EXPERIMENTAL] Added level 2
We added a functional version of Level 2, however it doesn’t look very
nice and has a few bugs that will be fixed in the next few days.

Important bug to fix:
1. Enemies do not respawn on retry.
2. Weird lighting (Fix is almost done, should be put out tomorrow in B5)

Thats the major bugs. A few GUI bugs need to be sorted, after that we
will add a custom main menu, but that is a while off, in July it will
be added (hopefully) Then this summer we’re going to move to Unity 4.6,
which will mean a lot better GUI because of there new system. Please
share this project with your friends so they can play this terrible
game!

——————————————————————————

www.youtube.com/user/wilsonkoder